requirements.txt
playwright==1.40.0
pandas==2.0.3
openpyxl==3.1.2
python-dotenv==1.0.0
schedule==1.2.0
jinja2==3.1.2
pytest==7.4.3
